requirements.txt
numpy==1.19.2
opencv==4.5.0
pillow==8.0.1
pytesseract==0.3.7
tesseract==0.1.3